Validate SpeciesList.when same as Observation
- Uses the same logic as Observation in order to prevent
User from making `when` `nil` by erasing the Year field in the form.
- Extract Observation `check_when` validation into concern
so that it can be Included in multiple models
See https://stackoverflow.com/questions/55583085/rails-accessing-a-shared-custom-validation-in-a-module-within-a-model
