+++ app/models/observation.rb
@@ -1169,9 +1169,10 @@ class Observation < AbstractModel
 
   protected
 
-  validate :check_requirements
-  def check_requirements # :nodoc:
-    check_when
+  include Validations
+
+  validate :check_requirements, :validate_when
+  def check_requirements
     check_where
     check_user
     check_coordinates
@@ -1187,47 +1188,6 @@ class Observation < AbstractModel
         errors.add(:when_str, :runtime_date_should_be_yyyymmdd.t)
       end
     end
-  end
-
-  def check_when
-    self.when ||= Time.zone.now
-    check_date && check_time && check_year
-  end
-
-  def check_date
-    return true unless self.when.is_a?(Date) && self.when > Time.zone.tomorrow
-
-    errors.add(:when, when_message("Time.zone.today=#{Time.zone.today}"))
-    errors.add(:when, :validate_observation_future_time.t)
-    false
-  end
-
-  def when_message(details = nil)
-    start = "self.when=#{self.when.class.name}:#{self.when}"
-    return start unless details
-
-    "#{start} #{details}"
-  end
-
-  def check_time
-    unless self.when.is_a?(Time) && self.when > Time.zone.now + 1.day
-      return true
-    end
-
-    # As of July 5, 2020 these statements appear to be unreachable
-    # because 'when' is a 'date' in the database.
-    errors.add(:when, when_message("Time.now=#{Time.zone.now + 6.hours}"))
-    errors.add(:when, :validate_observation_future_time.t)
-    false
-  end
-
-  def check_year
-    return true unless !self.when.respond_to?(:year) || self.when.year < 1500 ||
-                       self.when.year > (Time.zone.now + 1.day).year
-
-    errors.add(:when, when_message)
-    errors.add(:when, :validate_observation_invalid_year.t)
-    false
   end
 
   def check_where
